Report GCS

This function reports Glasgow Coma Scale (GCS). Individual elements as well as the sum of the score are important. Hence, the score is expressed in the form "GCS 9 = E2 V4 M3 at 2017-06-20 17:57:50".

reportgcs(eye = 4, verbal = 5, motor = 6,
  now = as.character(Sys.time()), simplify = TRUE)

Arguments

eye
Eye

verbal
Verbal

motor
Motor

now
Time

simplify
simplifying results or reporting full description

Examples

reportgcs()

#> [1] "GCS 15 = E4 V5 M6 at 2017-12-07 16:00:43 Normal or Minor brain injury suspected"

reportgcs(simplify = FALSE)

#> $summary
#> [1] "GCS 15 = E4 V5 M6 at 2017-12-07 16:00:43"
#> 
#> $analysis
#> [1] "Normal or Minor brain injury suspected"
#> 
#> $description
#>    ID  Class Score                       Value
#> 4  E4    Eye     4    Opens eyes spontaneously
#> 9  V5 Verbal     5 Oriented converses normally
#> 15 M6  Motor     6              Obeys commands
#> 

reportgcs(4, 5, 6, now = "2017-01-01 11:25") # custom time

#> [1] "GCS 15 = E4 V5 M6 at 2017-01-01 11:25 Normal or Minor brain injury suspected"

reportgcs(verbal = 2.5) # Error

#> Error in reportgcs(verbal = 2.5): Only integer is acceptable.

reportgcs(verbal = "erlk") # Error

#> Error in reportgcs(verbal = "erlk"): Only integer is acceptable.